% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeba.R
\name{adeba_density}
\alias{adeba_density}
\title{adeba_density}
\usage{
adeba_density(
  dat,
  range_log_alpha = c(-6, 1),
  vec_beta = c(0, 0.5, 1, 1.5, 2),
  discre_size1 = 14,
  discre_size2 = 7,
  alpha0 = 1,
  beta0 = 0.01,
  from = NA,
  to = NA,
  out = 0.2,
  m = 400
)
}
\arguments{
\item{dat}{A number}

\item{range_log_alpha}{hoge}

\item{vec_beta}{hoge}

\item{discre_size1}{hoge}

\item{discre_size2}{hoge}

\item{alpha0}{hoge}

\item{beta0}{hoge}

\item{from}{hoge}

\item{to}{hoge}

\item{out}{hoge}

\item{dat, }{hoge}

\item{n}{hoge}
}
\value{
huge
}
\description{
adeba_density
}
\examples{
set.seed(123); res <- adeba_density(rnorm(100))
plot(res, type="l")
curve(dnorm, -3, 3, add=TRUE, col="red")

}
