% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeba.R
\name{adeba_denstiy}
\alias{adeba_denstiy}
\title{Calculate Probability Density Using ADEBA Algorithm
same as density.adeba}
\usage{
adeba_denstiy(...)
}
\description{
Calculate Probability Density Using ADEBA Algorithm
same as density.adeba
}
