\name{badzupaR-package}
\alias{badzupaR-package}
\alias{badzupaR}
\docType{package}
\title{
  X-BADZUPA(A Cross-Validation and Bootstrap Approach for Detrital Zircon U-Pb age distributions)
}
\description{
  xbadzupaR is a comprehensive R package for detrital zircon U–Pb analysis. It features an optimal model selection algorithm with cross-validation, bootstrapping for uncertainty assessment, and Multi-Dimensional Scaling for distribution comparisons, making it a valuable tool for researchers studying sedimentary rock provenance and crustal history.
}
\details{
  \code{\link{density.adeba}{density.adeba}}
  \code{\link{Cv}{Cv}}
  \code{\link{Bootstrap}{Bootstrap}}
  \code{\link{plotBootstrapMDS}{plotBootstrapMDS}}
  \code{\link{getBootstrapMDS}{getBootstrapMDS}}
}
\author{
  Furukawa Tan
}
\references{
}
\keyword{  }
\seealso{
}
\examples{
  # \dontrun{
  #    ## Optional simple examples of the most important functions
  #    ## These can be in \dontrun{} and \donttest{} blocks.   
  # }
}
