% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{Cv}
\alias{Cv}
\title{R6 Class Bootstrap
text}
\description{
R6 Class Bootstrap
text

R6 Class Bootstrap
text
}
\examples{

## ------------------------------------------------
## Method `Cv$exeCVinAllAlgorithm`
## ------------------------------------------------

algorithms <- list(
    densityDefault = function(x) density(x),
    denstiySJ = function(x) density(x, bw = "SJ")
)
dat <- rnorm(100)
cvN <- 10
cv <- Cv$new(algorithms, dat, cvN)
res <- cv$exeCVinAllAlgorithm()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cv$new()}}
\item \href{#method-divideVector}{\code{Cv$divideVector()}}
\item \href{#method-logStar}{\code{Cv$logStar()}}
\item \href{#method-getA}{\code{Cv$getA()}}
\item \href{#method-rlcv}{\code{Cv$rlcv()}}
\item \href{#method-exeCV}{\code{Cv$exeCV()}}
\item \href{#method-exeCVinAllAlgorithm}{\code{Cv$exeCVinAllAlgorithm()}}
\item \href{#method-clone}{\code{Cv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$new(algorithms, dat, cvN, watcher = function(index, all) {
})}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-divideVector"></a>}}
\if{latex}{\out{\hypertarget{method-divideVector}{}}}
\subsection{Method \code{divideVector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$divideVector(N, vec)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logStar"></a>}}
\if{latex}{\out{\hypertarget{method-logStar}{}}}
\subsection{Method \code{logStar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$logStar(x, a)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getA"></a>}}
\if{latex}{\out{\hypertarget{method-getA}{}}}
\subsection{Method \code{getA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$getA(dat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rlcv"></a>}}
\if{latex}{\out{\hypertarget{method-rlcv}{}}}
\subsection{Method \code{rlcv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$rlcv(dat, dens, vdat)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exeCV"></a>}}
\if{latex}{\out{\hypertarget{method-exeCV}{}}}
\subsection{Method \code{exeCV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$exeCV(fn, seed = NULL, save = TRUE, algorithmName = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exeCVinAllAlgorithm"></a>}}
\if{latex}{\out{\hypertarget{method-exeCVinAllAlgorithm}{}}}
\subsection{Method \code{exeCVinAllAlgorithm()}}{
Change hair color.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$exeCVinAllAlgorithm(seed = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{New hair color.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{algorithms <- list(
    densityDefault = function(x) density(x),
    denstiySJ = function(x) density(x, bw = "SJ")
)
dat <- rnorm(100)
cvN <- 10
cv <- Cv$new(algorithms, dat, cvN)
res <- cv$exeCVinAllAlgorithm()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
