% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeba.R
\name{density.adeba}
\alias{density.adeba}
\title{Calculate Probability Density Using ADEBA Algorithm}
\usage{
\method{density}{adeba}(
  x,
  range_log_alpha = c(-6, 1),
  vec_beta = c(0, 0.5, 1, 1.5, 2),
  discre_size1 = 14,
  discre_size2 = 7,
  alpha0 = 1,
  beta0 = 0.01,
  from = NA,
  to = NA,
  out = 0.2,
  n = 400
)
}
\arguments{
\item{x}{A numeric vector representing the input data.}

\item{range_log_alpha}{A numeric vector of length 2 indicating the search range
for the parameter alpha. Increase the range if computation fails.}

\item{vec_beta}{A numeric vector containing candidate values for the parameter beta.}

\item{discre_size1}{The size of discretization for alpha values.}

\item{discre_size2}{Two values from discre_size1 are further discretized using discre_size2
around the maximum likelihood estimate of alpha to improve precision.}

\item{alpha0}{The initial value for the alpha parameter.}

\item{beta0}{The initial value for the beta parameter.}

\item{from}{The lower bound of the density estimation range.}

\item{to}{The upper bound of the density estimation range.}

\item{out}{The extra padding around the data range for drawing, if from and to are missing.}

\item{n}{The number of points to evaluate the density.}
}
\value{
A "density" object containing x (density values on the x-axis) and y (normalized density values).
}
\description{
This function calculates the probability density of input data using the ADEBA
(Automatic DEnsity Bandwidth Adjustment) algorithm. It employs a combination of
alpha and beta parameters to adjust the bandwidth for kernel density estimation.
}
\examples{
data <- rnorm(100)
density_result <- density.adeba(data)
plot(density_result, main = "Density Estimation")

}
