% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbadzupa.R
\name{getBestAlgorithmByCv}
\alias{getBestAlgorithmByCv}
\title{Get the best algorithm for density estimation using cross-validation.}
\usage{
getBestAlgorithmByCv(
  d,
  algorithms = list(densityADEBA = function(x) density.adeba(x, out = 0, n = 512),
    densityBOTEV = function(x) IsoplotR::kde(x, plot = FALSE))
)
}
\arguments{
\item{d}{The data for density estimation.}

\item{algorithms}{A list of density estimation functions to consider.}
}
\value{
function, the best density estimation algorithm.
}
\description{
This function uses cross-validation to determine the best density estimation
algorithm among the provided options.
}
\examples{
set.seed(1234); dat <- c(rnorm(200, 4), rnorm(50, 8, 2))
bestAlgorithm <- getBestAlgorithmByCv(dat)
plot(bestAlgorithm(dat))

}
