% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbadzupa.R
\name{getBootstrapResult}
\alias{getBootstrapResult}
\title{Get bootstrap results for a density estimation algorithm.}
\usage{
getBootstrapResult(dat, densFn, Nbootstrap = 400)
}
\arguments{
\item{dat}{The data for density estimation.}

\item{densFn}{The density estimation algorithm to use.}

\item{Nbootstrap}{The number of bootstrap samples to generate.}
}
\value{
A list containing bootstrap results and confidence intervals.
}
\description{
This function performs bootstrapping to obtain confidence intervals and other
information for a density estimation algorithm.
}
\examples{
set.seed(1234); dat <- c(rnorm(200, 4), rnorm(50, 8, 2))
bestAlgorithm <- getBestAlgorithmByCv(dat)
res <- getBootstrapResult(dat, bestAlgorithm)
plotBootstrapResult(res)

}
