% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.r
\name{plotBootstrapMDS}
\alias{plotBootstrapMDS}
\title{plotBootstrapMDS}
\usage{
plotBootstrapMDS(resBootstrapMDS, addPoints = FALSE)
}
\arguments{
\item{resBootstrapMDS}{A number}

\item{addPoints, }{hoge}
}
\description{
plotBootstrapMDS
}
\examples{
print("test")
}
