% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbadzupa.R
\name{plotBootstrapResult}
\alias{plotBootstrapResult}
\title{Plot bootstrap results for density estimation.}
\usage{
plotBootstrapResult(bootstrapRes, method = 1, showDetectedPeaks = TRUE)
}
\arguments{
\item{bootstrapRes}{The results of the bootstrap analysis.}

\item{method}{The plotting method (1 or 2).}

\item{showDetectedPeaks}{Whether to display detected peaks.}
}
\description{
This function plots bootstrap results, including confidence intervals and density
estimates, for a density estimation algorithm.
}
\examples{
set.seed(1234); dat <- c(rnorm(200, 4), rnorm(50, 8, 2))
bestAlgorithm <- getBestAlgorithmByCv(dat)
res <- getBootstrapResult(dat, bestAlgorithm)
plotBootstrapResult(res, method = 1)
plotBootstrapResult(res, method = 2)

}
