% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{Cv}
\alias{Cv}
\title{Cross Validation for xbadzupaR}
\description{
This R6 class provides cross-validation functionality for evaluating algorithms.
}
\examples{

## ------------------------------------------------
## Method `Cv$exeCVinAllAlgorithm`
## ------------------------------------------------

 algorithms <- list(
     denstiyBOTEV = function(x) IsoplotR::kde(x, plot=FALSE),
     densityADEBA = density.adeba
 )
 dat <- rnorm(100)
 cv <- Cv$new(algorithms, dat)
 cv$exeCVinAllAlgorithm()
 
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-cv-new}{\code{Cv$new()}}
\item \href{#method-cv-rlcv}{\code{Cv$rlcv()}}
\item \href{#method-cv-exeCV}{\code{Cv$exeCV()}}
\item \href{#method-cv-exeCVinAllAlgorithm}{\code{Cv$exeCVinAllAlgorithm()}}
\item \href{#method-cv-clone}{\code{Cv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv-new"></a>}}
\if{latex}{\out{\hypertarget{method-cv-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$new(
  algorithms,
  dat,
  watcher = function(index, all) {
 }
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algorithms}}{A list of algorithms to be evaluated.}

\item{\code{dat}}{The input data for cross-validation.}

\item{\code{watcher}}{A callback function to monitor progress.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Cv object.
Calculate Log-Likelihood with Bias for Cross-Validation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv-rlcv"></a>}}
\if{latex}{\out{\hypertarget{method-cv-rlcv}{}}}
\subsection{Method \code{rlcv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$rlcv(dat, dens, vdat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{A numeric vector representing the input data used for density estimation.}

\item{\code{dens}}{A "density" object representing the density estimation.}

\item{\code{vdat}}{A numeric vector representing the validation data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log-likelihood with added bias.

Execute cross-validation for a specific algorithm
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv-exeCV"></a>}}
\if{latex}{\out{\hypertarget{method-cv-exeCV}{}}}
\subsection{Method \code{exeCV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$exeCV(fn, seed = NULL, save = TRUE, algorithmName = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The algorithm function to be evaluated.}

\item{\code{seed}}{A seed value for reproducibility.}

\item{\code{save}}{A logical indicating whether to save results.}

\item{\code{algorithmName}}{The name of the algorithm.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing cross-validation results.
Execute cross-validation for all algorithms
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv-exeCVinAllAlgorithm"></a>}}
\if{latex}{\out{\hypertarget{method-cv-exeCVinAllAlgorithm}{}}}
\subsection{Method \code{exeCVinAllAlgorithm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$exeCVinAllAlgorithm(seed = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seed}}{A seed value for reproducibility.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing cross-validation results for all algorithms.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ algorithms <- list(
     denstiyBOTEV = function(x) IsoplotR::kde(x, plot=FALSE),
     densityADEBA = density.adeba
 )
 dat <- rnorm(100)
 cv <- Cv$new(algorithms, dat)
 cv$exeCVinAllAlgorithm()
 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv-clone"></a>}}
\if{latex}{\out{\hypertarget{method-cv-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
