% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.r
\name{getBootstrapMDS}
\alias{getBootstrapMDS}
\title{getBootstrapMDS}
\usage{
getBootstrapMDS(datList, ci = 0.9, Nbootstrap = 100)
}
\arguments{
\item{datList}{A number}

\item{ci, }{hoge}

\item{Nbootstrap}{hoge}
}
\description{
getBootstrapMDS
}
\examples{
print("test")

}
