% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.r
\name{plotBootstrapMDS}
\alias{plotBootstrapMDS}
\title{plotBootstrapMDS}
\usage{
plotBootstrapMDS(resBootstrapMDS, addPoints = FALSE)
}
\arguments{
\item{resBootstrapMDS}{A list containing MDS and ellipse coordinates obtained from `getBootstrapMDS`.}

\item{addPoints}{A logical value indicating whether to add bootstrapped points to the plot (default is FALSE).}
}
\description{
This function visualizes the Multidimensional Scaling (MDS) coordinates and confidence ellipses
generated by the `getBootstrapMDS` function. It provides a scatterplot of MDS coordinates and
optionally adds points from bootstrapped data along with ellipses representing confidence intervals.
}
\examples{
datList <- list(
  N1 = rnorm(200),
  N2 = rnorm(200),
  N3 = rnorm(200, 2, 1),
  N4 = rnorm(200, 2, 1),
  N5 = rnorm(200, 3, 1),
  N6 = rnorm(200, 3, 1)
  )
 res <- getBootstrapMDS(datList, Nbootstrap = 100)
 plotBootstrapMDS (res, addPoints=F)
}
